+++ internal/pkg/warewulfd/provision.go
@@ -71,7 +71,7 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 	var stage_file string
 
 	// TODO: when module version is upgraded to go1.18, should be 'any' type
-	var tmpl_data interface{}
+	var tmpl_data *templateVars
 
 	remoteNode, err := GetNodeOrSetDiscoverable(rinfo.hwaddr)
 	if err != nil && err != node.ErrNoUnconfigured {
@@ -91,13 +91,13 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 		wwlog.Error("%s (unknown/unconfigured node)", rinfo.hwaddr)
 		if rinfo.stage == "ipxe" {
 			stage_file = path.Join(conf.Paths.Sysconfdir, "/warewulf/ipxe/unconfigured.ipxe")
-			tmpl_data = templateVars{
+			tmpl_data = &templateVars{
 				Hwaddr: rinfo.hwaddr}
 		}
 
 	} else if rinfo.stage == "ipxe" {
 		stage_file = path.Join(conf.Paths.Sysconfdir, "warewulf/ipxe/"+remoteNode.Ipxe+".ipxe")
-		tmpl_data = templateVars{
+		tmpl_data = &templateVars{
 			Id:             remoteNode.Id(),
 			Cluster:        remoteNode.ClusterName,
 			Fqdn:           remoteNode.Id(),
@@ -181,7 +181,7 @@ func ProvisionSend(w http.ResponseWriter, req *http.Request) {
 			}
 		case "grub.cfg":
 			stage_file = path.Join(conf.Paths.Sysconfdir, "warewulf/grub/grub.cfg.ww")
-			tmpl_data = templateVars{
+			tmpl_data = &templateVars{
 				Id:             remoteNode.Id(),
 				Cluster:        remoteNode.ClusterName,
 				Fqdn:           remoteNode.Id(),